requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.4,<2.9.0
Pillow>=7.0.0,<7.1.0
django-hashid-field>=3.1.1,<3.2.0
celery==4.3.0
pinax-eventlog>=3.0.0,<3.1.0
redis>=3.4.1,<3.5.0

flake8>=3.7.9,<3.8.0
